martian: move proxyConn to separate file
